Refactor: dto 파일 올바른 위치로 이동
