--- src/main/java/com/jelly/zzirit/domain/timeDeal/dto/response/TimeDealResponseDTO.java
@@ -1,4 +1,0 @@
-package com.jelly.zzirit.domain.timeDeal.dto.response;
-
-public class TimeDealResponseDTO {
-}
